+++ src/dev/main.ts
@@ -1,10 +1,10 @@
 import { createApp } from "vue";
 import "./demo.css";
 import App from "./App.vue";
-import { VueForge } from "@/index";
+import { VueForgeCore } from "@/index";
 
 const app = createApp(App);
 
-app.use(VueForge);
+app.use(VueForgeCore);
 
 app.mount("#app");